% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junior2017.R
\docType{package}
\name{junior2017}
\alias{junior2017}
\alias{junior2017-package}
\title{Database of fiscal data for Brazilian Subnational Governments}
\description{
Database of fiscal data for Brazilian Subnational Governments
}
\details{
Package: junior2017

Type: Package

License: Public Domain (CC-0)
}
\references{
\url{https://github.com/fjuniorr/junior2017}
}
\author{
Francisco Alves \email{fjunior.alves.oliveira@gmail.com}
}
